0_Data_Setup_and_Clipping
This is a script pulling off earthstat and Mark's data for cocoa in Ghana and Cote d'Ivoire (two major producing countries listed on McVities info).  I have just clipped to countries but the scripts might be handy for anyone wanting to use a project and extend function on lots of maps at once, to extract country-specific clips from the EarthStat files, for instance.  Credits go to Tim Newbold and Fi Spooner, who helped me with some of the original functions used in this script.